requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
appdirs==1.4.4
    # via
    #   my-quants-app (setup.py)
    #   yfinance
appnope==0.1.3
    # via
    #   ipykernel
    #   ipython
    #   my-quants-app (setup.py)
asttokens==2.2.1
    # via
    #   my-quants-app (setup.py)
    #   stack-data
backcall==0.2.0
    # via
    #   ipython
    #   my-quants-app (setup.py)
beautifulsoup4==4.12.2
    # via
    #   feedfinder2
    #   my-quants-app (setup.py)
    #   newspaper3k
    #   yfinance
black==23.7.0
    # via my-quants-app (setup.py)
certifi==2023.7.22
    # via
    #   my-quants-app (setup.py)
    #   requests
cfgv==3.4.0
    # via
    #   my-quants-app (setup.py)
    #   pre-commit
charset-normalizer==3.2.0
    # via
    #   my-quants-app (setup.py)
    #   requests
click==8.1.6
    # via
    #   black
    #   my-quants-app (setup.py)
    #   nltk
comm==0.1.4
    # via
    #   ipykernel
    #   my-quants-app (setup.py)
contourpy==1.1.0
    # via
    #   matplotlib
    #   my-quants-app (setup.py)
cssselect==1.2.0
    # via
    #   my-quants-app (setup.py)
    #   newspaper3k
cycler==0.11.0
    # via
    #   matplotlib
    #   my-quants-app (setup.py)
debugpy==1.6.7.post1
    # via
    #   ipykernel
    #   my-quants-app (setup.py)
decorator==5.1.1
    # via
    #   ipython
    #   my-quants-app (setup.py)
distlib==0.3.7
    # via
    #   my-quants-app (setup.py)
    #   virtualenv
et-xmlfile==1.1.0
    # via
    #   my-quants-app (setup.py)
    #   openpyxl
executing==1.2.0
    # via
    #   my-quants-app (setup.py)
    #   stack-data
feedfinder2==0.0.4
    # via
    #   my-quants-app (setup.py)
    #   newspaper3k
feedparser==6.0.10
    # via
    #   my-quants-app (setup.py)
    #   newspaper3k
filelock==3.12.2
    # via
    #   my-quants-app (setup.py)
    #   tldextract
    #   virtualenv
fonttools==4.42.1
    # via
    #   matplotlib
    #   my-quants-app (setup.py)
frozendict==2.3.8
    # via
    #   my-quants-app (setup.py)
    #   yfinance
html5lib==1.1
    # via
    #   my-quants-app (setup.py)
    #   yfinance
identify==2.5.26
    # via
    #   my-quants-app (setup.py)
    #   pre-commit
idna==3.4
    # via
    #   my-quants-app (setup.py)
    #   requests
    #   tldextract
importlib-metadata==6.8.0
    # via
    #   jupyter-client
    #   my-quants-app (setup.py)
importlib-resources==6.0.1
    # via
    #   matplotlib
    #   my-quants-app (setup.py)
ipykernel==6.25.1
    # via my-quants-app (setup.py)
ipython==8.14.0
    # via
    #   ipykernel
    #   my-quants-app (setup.py)
jedi==0.19.0
    # via
    #   ipython
    #   my-quants-app (setup.py)
jieba3k==0.35.1
    # via
    #   my-quants-app (setup.py)
    #   newspaper3k
joblib==1.3.2
    # via
    #   my-quants-app (setup.py)
    #   nltk
jupyter-client==8.3.0
    # via
    #   ipykernel
    #   my-quants-app (setup.py)
jupyter-core==5.3.1
    # via
    #   ipykernel
    #   jupyter-client
    #   my-quants-app (setup.py)
kiwisolver==1.4.5
    # via
    #   matplotlib
    #   my-quants-app (setup.py)
lxml==4.9.3
    # via
    #   my-quants-app (setup.py)
    #   newspaper3k
    #   yfinance
matplotlib==3.7.2
    # via
    #   my-quants-app (setup.py)
    #   seaborn
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
    #   my-quants-app (setup.py)
multitasking==0.0.11
    # via
    #   my-quants-app (setup.py)
    #   yfinance
mypy-extensions==1.0.0
    # via
    #   black
    #   my-quants-app (setup.py)
nest-asyncio==1.5.7
    # via
    #   ipykernel
    #   my-quants-app (setup.py)
newspaper3k==0.2.8
    # via my-quants-app (setup.py)
nltk==3.8.1
    # via
    #   my-quants-app (setup.py)
    #   newspaper3k
nodeenv==1.8.0
    # via
    #   my-quants-app (setup.py)
    #   pre-commit
numpy==1.25.2
    # via
    #   contourpy
    #   matplotlib
    #   my-quants-app (setup.py)
    #   pandas
    #   pyarrow
    #   scipy
    #   seaborn
    #   yfinance
openpyxl==3.1.2
    # via my-quants-app (setup.py)
packaging==23.1
    # via
    #   black
    #   ipykernel
    #   matplotlib
    #   my-quants-app (setup.py)
pandas==2.0.3
    # via
    #   my-quants-app (setup.py)
    #   seaborn
    #   yfinance
parso==0.8.3
    # via
    #   jedi
    #   my-quants-app (setup.py)
pathspec==0.11.2
    # via
    #   black
    #   my-quants-app (setup.py)
pexpect==4.8.0
    # via
    #   ipython
    #   my-quants-app (setup.py)
pickleshare==0.7.5
    # via
    #   ipython
    #   my-quants-app (setup.py)
pillow==10.0.0
    # via
    #   matplotlib
    #   my-quants-app (setup.py)
    #   newspaper3k
platformdirs==3.10.0
    # via
    #   black
    #   jupyter-core
    #   my-quants-app (setup.py)
    #   virtualenv
polars==0.18.15
    # via my-quants-app (setup.py)
pre-commit==3.3.3
    # via my-quants-app (setup.py)
prompt-toolkit==3.0.39
    # via
    #   ipython
    #   my-quants-app (setup.py)
psutil==5.9.5
    # via
    #   ipykernel
    #   my-quants-app (setup.py)
ptyprocess==0.7.0
    # via
    #   my-quants-app (setup.py)
    #   pexpect
pure-eval==0.2.2
    # via
    #   my-quants-app (setup.py)
    #   stack-data
pyarrow==13.0.0
    # via my-quants-app (setup.py)
pygments==2.16.1
    # via
    #   ipython
    #   my-quants-app (setup.py)
pyparsing==3.0.9
    # via
    #   matplotlib
    #   my-quants-app (setup.py)
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   matplotlib
    #   my-quants-app (setup.py)
    #   newspaper3k
    #   pandas
pytz==2023.3
    # via
    #   my-quants-app (setup.py)
    #   pandas
    #   yfinance
pyyaml==6.0.1
    # via
    #   my-quants-app (setup.py)
    #   newspaper3k
    #   pre-commit
pyzmq==25.1.1
    # via
    #   ipykernel
    #   jupyter-client
    #   my-quants-app (setup.py)
regex==2023.8.8
    # via
    #   my-quants-app (setup.py)
    #   nltk
requests==2.31.0
    # via
    #   feedfinder2
    #   my-quants-app (setup.py)
    #   newspaper3k
    #   requests-file
    #   tldextract
    #   yfinance
requests-file==1.5.1
    # via
    #   my-quants-app (setup.py)
    #   tldextract
scipy==1.11.2
    # via my-quants-app (setup.py)
seaborn==0.12.2
    # via my-quants-app (setup.py)
sgmllib3k==1.0.0
    # via
    #   feedparser
    #   my-quants-app (setup.py)
simplejson==3.19.1
    # via my-quants-app (setup.py)
six==1.16.0
    # via
    #   asttokens
    #   feedfinder2
    #   html5lib
    #   my-quants-app (setup.py)
    #   python-dateutil
    #   requests-file
soupsieve==2.4.1
    # via
    #   beautifulsoup4
    #   my-quants-app (setup.py)
stack-data==0.6.2
    # via
    #   ipython
    #   my-quants-app (setup.py)
tinysegmenter==0.3
    # via
    #   my-quants-app (setup.py)
    #   newspaper3k
tldextract==3.4.4
    # via
    #   my-quants-app (setup.py)
    #   newspaper3k
tomli==2.0.1
    # via
    #   black
    #   my-quants-app (setup.py)
tornado==6.3.3
    # via
    #   ipykernel
    #   jupyter-client
    #   my-quants-app (setup.py)
tqdm==4.66.1
    # via
    #   my-quants-app (setup.py)
    #   nltk
traitlets==5.9.0
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   my-quants-app (setup.py)
typing-extensions==4.7.1
    # via
    #   black
    #   ipython
    #   my-quants-app (setup.py)
tzdata==2023.3
    # via
    #   my-quants-app (setup.py)
    #   pandas
urllib3==2.0.4
    # via
    #   my-quants-app (setup.py)
    #   requests
virtualenv==20.24.3
    # via
    #   my-quants-app (setup.py)
    #   pre-commit
wcwidth==0.2.6
    # via
    #   my-quants-app (setup.py)
    #   prompt-toolkit
webencodings==0.5.1
    # via
    #   html5lib
    #   my-quants-app (setup.py)
yfinance==0.2.28
    # via my-quants-app (setup.py)
zipp==3.16.2
    # via
    #   importlib-metadata
    #   importlib-resources
    #   my-quants-app (setup.py)

# The following packages are considered to be unsafe in a requirements file:
# setuptools
